Séparation responsabilités solo / multi
